+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 🧩 Addon Updates (2026-06-01)
 
-**Mythic Dungeon Teleports** — v2.1.5  
+**Mythic Dungeon Teleports** — v2.1.6  
 
 **Changes:**  
-- None
+- Updated Keystone Module to display realm next to player name. This is to prevent confusion with the same character name over different servers
+- Keystone Module is now fully localised in all languages available in Mythic Dungeon Teleports
 
 **Fixes:**  
 - Update various translations for reported bug #25